chore(utils): Move fetch functions to utils/api/client.ts
